// local-hobbies-frontend/src/types/user.ts
export enum UserGenderMatched {
    Man = 'Man',
    Nonbinary = 'Nonbinary',
    Woman = 'Woman',
}
// ...
export interface UserResponse {
    username: string;
    email: string;
    name: string;
    /** ISO 8601 date */
    birthDate: string;
    genderDisplayed: string | null;
    bio: string | null;
    locationPoint: GeoJsonPoint;
    locationApproximate: string;
    publicContactInfo: string;
    profilePhotoUrl: string | null;
    genderMatched: UserGenderMatched;
    showAge: boolean;
    showGenderDisplayed: boolean;
}